refactor: move Likelihood-related code to new `likelihoods` module
